norminette adjust

+++ src/stack_functions.c
@@ -6,13 +6,13 @@
 /*   By: rsiqueir <rsiqueir@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/11 02:23:50 by coder             #+#    #+#             */
-/*   Updated: 2022/09/12 21:32:27 by rsiqueir         ###   ########.fr       */
+/*   Updated: 2022/09/14 22:35:28 by rsiqueir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-# include "../pushswap.h"
+#include "../pushswap.h"
 
-void    move_memmory_foward(t_stack *stack, size_t size)
+void	move_memmory_foward(t_stack *stack, size_t size)
 {
 	if (size == 0)
 		return ;
@@ -25,7 +25,7 @@ void    move_memmory_foward(t_stack *stack, size_t size)
 
 void	move_memmory_back(t_stack *stack, size_t size)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (size == 0)
@@ -39,19 +39,19 @@ void	move_memmory_back(t_stack *stack, size_t size)
 
 void	swap(t_stack *a, t_stack *b)
 {
-	t_stack tmp;
+	t_stack	tmp;
 
 	tmp = *a;
 	*a = *b;
 	*b = tmp;
 }
 
-int find_min_number(t_stack *stack, size_t size)
+int	find_min_number(t_stack *stack, size_t size)
 {
-    int i;
-	int	number;
+	size_t	i;
+	int		number;
 
-    i = 0;
+	i = 0;
 	number = stack[0].content;
 	while (++i < size)
 	{
@@ -61,7 +61,7 @@ int find_min_number(t_stack *stack, size_t size)
 	return (number);
 }
 
-void clear_stack(t_push *ps)
+void	clear_stack(t_push *ps)
 {
 	free(ps->stack_a);
 	free(ps->stack_b);
